requirements.txt
# cat requirements.txt 
flask
openstacksdk

